django-vue3-admin-master/backend/requirements.txt
Django==5.2.0
django-comment-migrate==0.1.7
django-cors-headers==4.4.0
django-filter==24.2
django-ranged-response==0.2.0
djangorestframework==3.16.0
django-restql==0.15.4
django-simple-captcha==0.6.0
django-timezone-field==7.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.7
pymysql==1.1.0
pypinyin==0.51.0
ua-parser==0.18.0
pyparsing==3.1.2
openpyxl==3.1.5
requests==2.32.4
typing-extensions==4.12.2
tzlocal==5.2
channels==4.1.0
channels-redis==4.2.0
user-agents==2.2.0
whitenoise==6.7.0
psycopg2-binary==2.9.11
uvicorn==0.30.3
gunicorn==23.0.0
gevent==25.9.1
Pillow>=11.0.0
django-celery-beat==2.8.1
django-celery-results==2.5.1
django-redis==5.4.0
tenant-schemas-celery==3.0.0
oss2==2.19.1
cos-python-sdk-v5==1.9.37